+++ src/ambition/version_details.cc
@@ -13,6 +13,7 @@ std::vector<VersionUpdates> versionDetails = {
         .heading = _("Added"),
         .items = {
           _("Automatically copy the map ID to the clipboard when opening the in-game menu, allowing for easy copy-pasting of the map ID."),
+          _("Automatically paste into the join address field, allowing for easy joining of games when using the \"Enter Address\" option."),
           _("Added a What's New screen, shown on the first startup of each new version."),
         },
       },