+++ CHANGELOG.md
@@ -14,6 +14,8 @@ speaking, considered to be the player's experience.
 
 - Automatically copy the map ID to the clipboard when opening the in-game menu,
   allowing for easy copy-pasting of the map ID.
+- Automatically paste into the join address field, allowing for easy joining of
+  games when using the "Enter Address" option.
 - Added a What's New screen, shown on the first startup of each new version.
 
 ### Fixed